Cockpitコンポーネント作成
