+++ src/container/App.css
@@ -1,33 +1,3 @@
 .App {
   text-align: center;
 }
-
-.red {
-  color: red;
-}
-
-.bold {
-  font-weight: bold;
-}
-
-.App button {
-  background-color: green;
-  color: white;
-  font: inherit;
-  border: 10px solid black;
-  padding: 8px;
-  cursor: pointer;
-}
-
-.App button:hover {
-  background-color: lightgreen;
-  color: black;
-}
-
-.App button.Red {
-  background-color: red;
-}
-
-.App button.Red:hover {
-  background-color: salmon;
-}
